initialLoad from a specific given revision
This seems smarter for concurrent start up, where an already up node
could already be in the process of actively writing while another one is
the process of coming up.

Also note how in init() we now FIRST update the revAwaiter and THEN
start the Watcher - that makes more sense.